lapd-crime-app/requirements.txt
streamlit
pandas
plotly
pyarrow
requests
databricks-sdk>=0.29.0
databricks-sql-connector
